extract OgrreVersionDialog.tsx into its own component, ensure we display new info
