+++ src/components/Header/Header.tsx
@@ -6,77 +6,27 @@ import { useUserContext } from "../../usercontext";
 import { changeTeam, fetchTeams, getOgrreVersion } from "../../services/app.service";
 import { ChangeTeamResponse } from "../../types";
 import ChangeTeamDialog from "./ChangeTeamDialog";
+import OgrreVersionDialog, { OgrreVersionInfo } from "./OgrreVersionDialog";
 import {
-  Alert,
   Avatar,
-  Box,
   Button,
-  CircularProgress,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogTitle,
   Divider,
   IconButton,
   ListItemIcon,
   Menu,
   MenuItem,
-  Stack,
   Tab,
   Tabs,
-  Typography,
 } from "@mui/material";
 import { logout, callAPI } from "../../util";
-import CloseIcon from "@mui/icons-material/Close";
 import InfoOutlinedIcon from "@mui/icons-material/InfoOutlined";
 import Logout from "@mui/icons-material/Logout";
 import OpenInNewIcon from "@mui/icons-material/OpenInNew";
 import GroupsOutlinedIcon from "@mui/icons-material/GroupsOutlined";
 
-interface PackageVersionInfo {
-  name: string;
-  version?: string;
-  commit?: string;
-  source_url?: string;
-  requested_revision?: string;
-}
-
-interface OgrreVersionInfo {
-  packages: PackageVersionInfo[];
-}
-
-const formatPackageName = (packageName: string) => {
-  if (packageName === "ogrre_data_cleaning") return "ogrre_data_cleaning";
-  if (packageName === "orphaned-wells-ui-server") return "orphaned-wells-ui-server";
-  return packageName;
-};
-
 const getApiErrorMessage = (error: any, fallback: string) => {
   if (typeof error === "string") return error;
   return error?.message || error?.detail || fallback;
-};
-
-const VersionMetadataLine = ({ label, value }: { label: string; value?: string }) => {
-  if (!value) return null;
-  return (
-    <Stack direction="row" spacing={1.5} sx={{ alignItems: "flex-start" }}>
-      <Typography sx={{ color: "#6B7280", fontSize: "13px", minWidth: "92px" }}>
-        {label}
-      </Typography>
-      <Box
-        component="span"
-        sx={{
-          color: "#111827",
-          fontFamily: "ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, \"Liberation Mono\", \"Courier New\", monospace",
-          fontSize: "12px",
-          lineHeight: 1.6,
-          overflowWrap: "anywhere",
-        }}
-      >
-        {value}
-      </Box>
-    </Stack>
-  );
 };
 
 const Header = (props: any) => {
@@ -294,85 +244,13 @@ const Header = (props: any) => {
         onClose={handleCloseTeamDialog}
         onChangeTeam={handleChangeTeam}
       />
-      <Dialog
+      <OgrreVersionDialog
         open={versionDialogOpen}
+        versionInfo={versionInfo}
+        loading={versionLoading}
+        error={versionError}
         onClose={() => setVersionDialogOpen(false)}
-        fullWidth
-        maxWidth="sm"
-        PaperProps={{
-          sx: {
-            borderRadius: "12px",
-            border: "1px solid #E5E7EB",
-          },
-        }}
-      >
-        <DialogTitle
-          sx={{
-            borderBottom: "1px solid #EEF2F7",
-            pr: 6,
-          }}
-        >
-          <Typography component="div" variant="h6" sx={{ fontWeight: 700 }}>
-            OGRRE Version
-          </Typography>
-          <Typography component="div" sx={{ color: "#6B7280", fontSize: "13px", mt: 0.5 }}>
-            Backend package metadata currently reported by the server.
-          </Typography>
-        </DialogTitle>
-        <IconButton
-          aria-label="close"
-          onClick={() => setVersionDialogOpen(false)}
-          sx={{ position: "absolute", right: 8, top: 8 }}
-        >
-          <CloseIcon />
-        </IconButton>
-        <DialogContent dividers sx={{ p: 2.5 }}>
-          {versionLoading ? (
-            <Stack direction="row" spacing={1.5} alignItems="center">
-              <CircularProgress size={22} />
-              <Typography sx={{ color: "#4B5563", fontSize: "14px" }}>
-                Loading version information...
-              </Typography>
-            </Stack>
-          ) : versionError ? (
-            <Alert severity="error">{versionError}</Alert>
-          ) : versionInfo?.packages?.length ? (
-            <Stack spacing={1.5}>
-              {versionInfo.packages.map((packageInfo) => (
-                <Box
-                  key={packageInfo.name}
-                  sx={{
-                    border: "1px solid #E5E7EB",
-                    borderRadius: "8px",
-                    backgroundColor: "#FFFFFF",
-                    p: 1.5,
-                  }}
-                >
-                  <Typography sx={{ fontWeight: 700, color: "#111827", mb: 1 }}>
-                    {formatPackageName(packageInfo.name)}
-                  </Typography>
-                  <Stack spacing={0.75}>
-                    <VersionMetadataLine label="Version" value={packageInfo.version || "Unknown"} />
-                    <VersionMetadataLine label="Commit" value={packageInfo.commit} />
-                    <VersionMetadataLine
-                      label="Requested"
-                      value={packageInfo.requested_revision !== packageInfo.commit ? packageInfo.requested_revision : undefined}
-                    />
-                    <VersionMetadataLine label="Source" value={packageInfo.source_url} />
-                  </Stack>
-                </Box>
-              ))}
-            </Stack>
-          ) : (
-            <Typography sx={{ color: "#4B5563", fontSize: "14px" }}>
-              No version information was returned.
-            </Typography>
-          )}
-        </DialogContent>
-        <DialogActions sx={{ px: 2, py: 1.5 }}>
-          <Button onClick={() => setVersionDialogOpen(false)}>Close</Button>
-        </DialogActions>
-      </Dialog>
+      />
     </div>
   );
 };
